requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2.20
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.4.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.15
